etcdserver/etcdhttp: break apart http.go
